создание массива из объектов из фото разбито на модули
